test(ui): split call-to-arms 320 dp pin to own file (Refs #2870)
The previous commit pushed `dialogs_320dp_min_viewport_test.dart` to
1078 non-comment lines, breaching the
`repo.dart_file_non_comment_line_size` 1000-line budget and failing
`Quality (lint, analyze, observer)` on PR #3069.

Extract the new `CallToArmsDialogueOverlay` group into a sibling file
`app/test/call_to_arms_dialogue_overlay_320dp_min_viewport_test.dart`
(mirrors the existing per-category split convention used by
`move_dialogs_320dp_min_viewport_test.dart`,
`train_dialogs_320dp_min_viewport_test.dart`, etc., each duplicating the
small `_kMinViewport` / `_kWideRegressionViewport` / `_pumpDialogAtSize`
private helpers).

After the split:

- `dialogs_320dp_min_viewport_test.dart`: 1176 total / 960 non-comment
  lines (back under the 1000 cap).
- `call_to_arms_dialogue_overlay_320dp_min_viewport_test.dart`: 225
  total / 150 non-comment lines.
- Host file's preamble updated to note the sibling file; unused
  `call_to_arms_dialogue_overlay.dart` + `CallToArmsPending` imports
  removed.

No SPEC or behavior changes; the three call-to-arms 320 dp ACs are
unchanged.

Verification:
- `flutter test test/dialogs_320dp_min_viewport_test.dart test/call_to_arms_dialogue_overlay_320dp_min_viewport_test.dart` -> 32/32 pass
- `flutter test test/call_to_arms_dialogue_overlay_dark_chrome_test.dart` -> 5/5 pass
- `flutter analyze` on both files -> no issues
- `CT_REPO_LINT_INCLUDE_APP=true dart run tool/ct_repo_lint.dart` -> all 53 rules pass (was failing `repo.dart_file_non_comment_line_size` before this commit).
